move files inside backend directory
